Move thread functions to their own header file
