+++ src/c/common.h
@@ -177,64 +177,4 @@ char *join_strings(const char **strings, size_t count, const char *delim) {
     return result;
 }
 
-// =============================
-// THREAD SYNCHRONIZATION HELPERS
-// =============================
-
-static inline void ctx_lock() {
-    pthread_mutex_lock(&ctx->mutex);
-}
-
-static inline void ctx_unlock() {
-    pthread_mutex_unlock(&ctx->mutex);
-}
-
-/*
- * Transition the thread context to a new state while holding the mutex.
- * This ensures atomic state transitions with proper synchronization.
- */
-static inline void ctx_set_state(ThreadState new_state) {
-    ctx->state = new_state;
-}
-
-/*
- * Signal the main thread to wake up and check for work.
- */
-static inline void ctx_signal_main() {
-    pthread_cond_signal(&ctx->cond);
-}
-
-/*
- * Wait for the state to change from the current state.
- * Must be called with mutex locked; returns with mutex still locked.
- */
-static inline void ctx_wait_for_state_change(ThreadState expected_state) {
-    while (ctx->state == expected_state) {
-        pthread_cond_wait(&ctx->cond, &ctx->mutex);
-    }
-}
-
-/*
- * Request that the main thread execute a directive.
- * Blocks until the directive completes (or signals early completion).
- * Returns the directive result code.
- */
-int ctx_request_main_execution(const char *directive, size_t dir_argc, const char **dir_argv);
-
-/*
- * Signal that the current directive has completed early and will continue
- * running asynchronously (e.g., blocking in a runloop).
- * This releases the directive thread to continue processing.
- * Must be called from main thread while executing a directive.
- */
-void ctx_signal_early_completion(ThreadState new_state);
-
-/*
- * Check if the current thread is the main thread by examining the context state.
- * Returns 1 if main thread is available for directive execution, 0 otherwise.
- */
-static inline int ctx_main_thread_available() {
-    return ctx->state == STATE_WAITING;
-}
-
 #endif
